Add place counter, revise lookupNode and getSuggestions
Counting places so we can make sure that the nodes are added correctly. Changed lookupNode to be a bit more relaxed and also return partial matches. Because this is needed in getSuggestions to actually get suggestions you need a node to start from. Therefore we are content with a partial match